Rename recyclerviewadapters package to adapters
