+++ app/src/main/java/com/shapeapp/shape/ui/chat/ChatFragment.kt
@@ -10,7 +10,7 @@ import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.shapeapp.shape.R
-import com.shapeapp.shape.ui.recyclerviews.recyclerviewadapters.MessagesFullRecyclerViewAdapter
+import com.shapeapp.shape.ui.recyclerviews.adapters.MessagesFullRecyclerViewAdapter
 import kotlinx.android.synthetic.main.fragment_chat.*
