+++ app/src/main/java/com/shapeapp/shape/ui/dialogs/MyPublicSharesBSD.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.shapeapp.shape.R
-import com.shapeapp.shape.ui.recyclerviews.recyclerviewadapters.SmallCardRecyclerViewAdapter
+import com.shapeapp.shape.ui.recyclerviews.adapters.SmallCardRecyclerViewAdapter
 import com.shapeapp.shape.repositories.CardsRepository
 import kotlinx.android.synthetic.main.bsd_my_public_shares.*
